Suggested changes implemented
